production_requirements.txt
Django==1.8.5
django-allauth==0.23.0
django-debug-toolbar==1.3.2
django-enumfield==1.2.1
django-formtools==1.0
oauthlib==1.0.3
python-openid==2.2.5
requests==2.8.0
requests-oauthlib==0.5.0
sqlparse==0.1.16
wheel==0.24.0
